Introduced Logging Mechanism:
As a result of the new Logging Mechanism, I've removed the use of RWST.
I naively thought I'd use the "writer" portion of RWST to do logging,
but it's unneccessary. I wanted to use file based logging for the test
suite, so I could see why things went wrong easier. However, this
implies the use of IO. Since IO is already part of the transformer stack
I just decided to scrub RWST, as I don't need the Writer portion, and
decided to use StateT.

The basic gist of Logging is that it opens a file at the start of
execution and will write data to it as needed.