Fixed `onRegister` function:
The `onRegister` function "lifts" a `Byte -> Byte -> Byte` binary
operation into the `Register` type. Unfortunately, it's usage does not
account for the LHS and RHS to be in a specific order (commutative? associative?).
Since we are only using this function when using modifying lenses, the
value we want to modify is actually the second parameter. This caused
some unexpected behavior i.e. ASL -> shifted 1 by the A register rather
than the other way around.